+++ Project/RoomRentalProject/DBL/User_Service/UserService/UserActionClass/EditUser.cs
@@ -6,7 +6,7 @@
         public string? name { get; set; }
         public string? password { get; set; }
         public string? email { get; set; }
-        public int userRoleId { get; set; }
+        public short userRoleId { get; set; }
         public string? phone { get; set; }
         public string? address { get; set; }
     }
